pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Ruslan Akhmetshin <sccrsccr1@gmail.com>"]
packages = [{ include = "task_manager" }]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
Django = "^4.1"
gunicorn = "^20.1.0"
django-bootstrap4 = "^22.1"
whitenoise = "^6.2.0"
psycopg2 = "^2.9.3"
pytest = "^7.1.2"
django-filter = "^22.1"
django-bootstrap-form = "^3.4"
rollbar = "^0.16.3"

[tool.poetry.dev-dependencies]
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"
isort = "^5.10.1"
black = "^22.6.0"
flake8 = "^5.0.4"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
